moved SetupTestCase into tests.test_bin
